Title: 세 수의 최대 곱, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240530/세 수의 최대 곱/maximum-product-of-three-numbers.cpp
@@ -17,7 +17,7 @@ int main() {
         {
             for(int k=j+1;k<n;k++)
             {
-                int mul=arr[i]*arr[j]*arr[k];
+                long long mul=arr[i]*arr[j]*arr[k];
                 if(mul>0)
                 {
                     maxi=max(maxi,mul);
